fai_m10_v3/requirements.txt
# Core dependencies for L3_M10.3: Managing Financial Knowledge Base Drift

# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Service-specific dependencies (auto-detected from script Section 4)
# OpenAI (primary): Embeddings API for drift detection
openai==1.3.5

# Pinecone (secondary): Vector database for knowledge storage
pinecone-client==2.2.4

# PostgreSQL for audit trails (SOX compliance)
psycopg2-binary==2.9.9

# Web scraping for regulatory monitoring (FASB, SEC, AICPA)
beautifulsoup4==4.12.2
requests==2.31.0

# Scientific computing for embeddings and similarity
scikit-learn==1.3.2
numpy==1.26.2

# Testing framework
pytest==7.4.3
httpx==0.25.1
pytest-cov==4.1.0

# Development and utilities
python-dotenv==1.0.0
